guardian-analyzer/pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
name = "guardian_analyzer"
version = "2.2.355"
description = "Guardian Analyzer package"
authors = ["Naufil <naufil.asar23@spit.ac.in>"]
license = "MIT"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
keywords = ["presidio_analyzer"]
urls = {Homepage = "https://github.com/SafeGuardians/presidio"}
readme = "README.md"
include = ["conf/*",]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
spacy= ">=3.4.4, <4.0.0"
regex = "*"
tldextract = "*"
pyyaml = "*"
phonenumbers = ">=8.12,<9.0.0"
flask = { version = ">=1.1", optional = true }
spacy_huggingface_pipelines = { version = "*", optional = true }
stanza = { version = "*", optional = true }
spacy_stanza = { version = "*", optional = true }
azure-ai-textanalytics = { version = "*", optional = true }
azure-core = { version = "*", optional = true }
transformers = { version = "*", optional = true }
huggingface_hub = { version = "*", optional = true }
pymupdf = "^1.25.0"
typing-extensions = ">=4.0.0"
pillow = "^11.0.0"
opencv-python = "^4.10.0.84"
pytesseract = "^0.3.13"

[tool.poetry.extras]
server = ["flask"]
transformers = [
    "transformers",
    "huggingface_hub",
    "spacy_huggingface_pipelines"]
stanza = [
    "stanza",
    "spacy_stanza",
]
azure-ai-language = [
    "azure-ai-textanalytics",
    "azure-core",
]

[tool.poetry.group.dev.dependencies]
pip = "*"
ruff = "*"
pytest = "*"
pytest-mock = "*"
python-dotenv = "*"
pre_commit = "*"
